Clean up signals and recievers
