+++ gcampus/auth/signals.py
@@ -13,8 +13,9 @@
 #  You should have received a copy of the GNU Affero General Public License
 #  along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-__all__ = ["token_user_logged_in"]
+__all__ = ["token_user_logged_in", "course_updated"]
 
 from django.dispatch import Signal
 
 token_user_logged_in = Signal()
+course_updated = Signal()
